feat: directly assign System.Uri and System.Version when deep cloning
Since System.Uri and System.Version are immutable, these can be directly assigned even if deep cloning is enabled
